winai-jobs/requirements.txt
psycopg2-binary
openai
pandas
schedule
python-dotenv
sqlalchemy
